refactor(components): "modularize the monolith" and extract bloated page components & inline logic
- Rename components/card to Card & update imports
- Extract IngoLogo & DRY navbar & footer & cleanup comments
