+++ components/Card/Bulletin.js
