+++ components/Card/Thesis.js
