+++ components/Navbar.js
@@ -6,6 +6,16 @@ import { useEffect, useState } from 'react';
 import Link from 'next/link';
 import { useRouter } from 'next/router';
 import { usePrefetcher } from './Prefetcher';
+import IngoLogo from './IngoLogo';
+
+const NAV_LINKS = [
+  { href: '/blog', label: 'blog' },
+  { href: '/bulletin', label: 'bulletin' },
+  { href: '/thesis', label: 'thesis' },
+  { href: '/awards', label: 'awards' },
+  { href: '/gallery', label: 'gallery' },
+  { href: '/about', label: 'about' },
+];
 
 const Navbar = () => {
   const router = useRouter();
@@ -99,61 +109,17 @@ const Navbar = () => {
               }}
               className="w-10/12 sm:w-6/12 min-h-screen bg-[#161a22] flex flex-col justify-center gap-2 px-4"
             >
-              <Link href="/blog">
-                <Button
-                  onClick={(e) => setSideMenuOpen(false)}
-                  variant="text"
-                  className={isActivePath('/blog') ? 'text-header-color' : 'text-button-color'}
-                >
-                  blog
-                </Button>
-              </Link>
-              <Link href="/bulletin">
-                <Button
-                  onClick={(e) => setSideMenuOpen(false)}
-                  variant="text"
-                  className={isActivePath('/bulletin') ? 'text-header-color' : 'text-button-color'}
-                >
-                  bulletin
-                </Button>
-              </Link>
-              <Link href="/thesis">
-                <Button
-                  onClick={(e) => setSideMenuOpen(false)}
-                  variant="text"
-                  className={isActivePath('/thesis') ? 'text-header-color' : 'text-button-color'}
-                >
-                  thesis
-                </Button>
-              </Link>
-              <Link href="/awards">
-                <Button
-                  onClick={(e) => setSideMenuOpen(false)}
-                  variant="text"
-                  className={isActivePath('/awards') ? 'text-header-color' : 'text-button-color'}
-                >
-                  awards
-                </Button>
-              </Link>
-              <Link href="/gallery">
-                <Button
-                  onClick={(e) => setSideMenuOpen(false)}
-                  variant="text"
-                  className={isActivePath('/gallery') ? 'text-header-color' : 'text-button-color'}
-                >
-                  gallery
-                </Button>
-              </Link>
-
-              <Link href="/about">
-                <Button
-                  onClick={(e) => setSideMenuOpen(false)}
-                  variant="text"
-                  className={isActivePath('/about') ? 'text-header-color' : 'text-button-color'}
-                >
-                  about
-                </Button>
-              </Link>
+              {NAV_LINKS.map(({ href, label }) => (
+                <Link key={href} href={href}>
+                  <Button
+                    onClick={() => setSideMenuOpen(false)}
+                    variant="text"
+                    className={isActivePath(href) ? 'text-header-color' : 'text-button-color'}
+                  >
+                    {label}
+                  </Button>
+                </Link>
+              ))}
             </motion.div>
           </motion.div>
         )}
@@ -253,66 +219,19 @@ const Navbar = () => {
               </IconButton>
             </div>
             <Link href="/">
-              <p className="text-2xl font-bold text-transparent cursor-pointer">
-                <motion.span
-                  animate={{
-                    backgroundPosition: [
-                      '0% 0%',
-                      '100% 0%',
-                      '100% 100%',
-                      '0% 100%',
-                      '0% 0%',
-                    ],
-                  }}
-                  transition={{
-                    duration: 7,
-                    ease: 'linear',
-                    repeat: Infinity,
-                  }}
-                  style={{
-                    backgroundSize: '1000px 1000px',
-                    backgroundColor: 'rgb(255, 50, 6)',
-                    backgroundImage:
-                      'radial-gradient(at 0% 100%, rgb(244, 63, 94) 0, transparent 50%), radial-gradient(at 90% 0%, rgb(100, 50, 85) 0, transparent 50%), radial-gradient(at 100% 100%, rgb(217, 70, 239) 0, transparent 50%), radial-gradient(at 0% 0%, rgb(249, 115, 22) 0, transparent 58%)',
-                  }}
-                  className="bg-clip-text bg-transparent"
-                >
-                  ingo
-                </motion.span>
-              </p>
+              <div className="cursor-pointer">
+                <IngoLogo />
+              </div>
             </Link>
           </div>
           <div className="lg:flex gap-2 hidden">
-            <Link href="/blog" scroll={false}>
-              <Button className={`font-black ${isActivePath('/blog') ? 'text-header-color' : 'text-nav-color'}`} variant="text">
-                blog
-              </Button>
-            </Link>
-            <Link href="/bulletin" scroll={false}>
-              <Button className={`font-black ${isActivePath('/bulletin') ? 'text-header-color' : 'text-nav-color'}`} variant="text">
-                bulletin
-              </Button>
-            </Link>
-            <Link href="/thesis" scroll={false}>
-              <Button className={`font-black ${isActivePath('/thesis') ? 'text-header-color' : 'text-nav-color'}`} variant="text">
-                thesis
-              </Button>
-            </Link>
-            <Link href="/awards" scroll={false}>
-              <Button className={`font-black ${isActivePath('/awards') ? 'text-header-color' : 'text-nav-color'}`} variant="text">
-                awards
-              </Button>
-            </Link>
-            <Link href="/gallery" scroll={false}>
-              <Button className={`font-black ${isActivePath('/gallery') ? 'text-header-color' : 'text-nav-color'}`} variant="text">
-                gallery
-              </Button>
-            </Link>
-            <Link href="/about" scroll={false}>
-              <Button className={`font-black ${isActivePath('/about') ? 'text-header-color' : 'text-nav-color'}`} variant="text">
-                about
-              </Button>
-            </Link>
+            {NAV_LINKS.map(({ href, label }) => (
+              <Link key={href} href={href} scroll={false}>
+                <Button className={`font-black ${isActivePath(href) ? 'text-header-color' : 'text-nav-color'}`} variant="text">
+                  {label}
+                </Button>
+              </Link>
+            ))}
             <Tooltip
               content="You can press CTRL + K to open search"
               placement="bottom-end"
