+++ pages/gallery/index.js
